+++ trixnity-client/trixnity-client-media-okio/src/commonTest/kotlin/net/folivo/trixnity/client/media/okio/OkioMediaStoreTest.kt
@@ -1,7 +1,11 @@
 package net.folivo.trixnity.client.media.okio
 
+import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.test.runTest
 import net.folivo.trixnity.utils.toByteArray
@@ -16,8 +20,10 @@ class OkioMediaStoreTest {
 
     lateinit var cut: OkioMediaStore
     lateinit var fileSystem: FakeFileSystem
+    lateinit var coroutineScope: CoroutineScope
 
     private val basePath = "/data/media".toPath()
+    private val tmpPath = basePath.resolve("tmp")
     private val file1 = "/data/media/file1".toPath()
     private val file2 = "/data/media/file2".toPath()
 
@@ -25,40 +31,43 @@ class OkioMediaStoreTest {
     fun beforeTest() {
         fileSystem = FakeFileSystem()
         cut = OkioMediaStore(basePath, fileSystem)
+        coroutineScope = CoroutineScope(Dispatchers.Default)
     }
 
     @AfterTest
     fun afterTest() {
         fileSystem.checkNoOpenFiles()
+        coroutineScope.cancel()
     }
 
     @Test
     fun shouldInit() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         fileSystem.exists(basePath) shouldBe true
-        cut.init() // should not fail
+        fileSystem.exists(tmpPath) shouldBe true
+        cut.init(coroutineScope) // should not fail
     }
 
     @Test
     fun shouldDeleteAll() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         fileSystem.write(file1) {}
         fileSystem.write(file2) {}
-        fileSystem.listOrNull(basePath)?.size shouldBe 3
+        fileSystem.listOrNull(basePath)?.size shouldBe 4
         cut.deleteAll()
-        fileSystem.listOrNull(basePath)?.size shouldBe 1
+        fileSystem.listOrNull(basePath)?.size shouldBe 2
     }
 
     @Test
     fun shouldAddMedia() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         cut.addMedia("url1", "hi".encodeToByteArray().toByteArrayFlow())
         fileSystem.read(basePath.resolve("K5pAaUF5iDoN1BsrFr4kJ0bP8ayM_Q_ftEtyeb_FY2I=")) { readUtf8() } shouldBe "hi"
     }
 
     @Test
     fun shouldNotAddMediaOnException() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         val file = MutableSharedFlow<ByteArray>()
         val job = async {
             cut.addMedia("url1", file)
@@ -70,7 +79,7 @@ class OkioMediaStoreTest {
 
     @Test
     fun shouldGetMedia() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         fileSystem.write(basePath.resolve("K5pAaUF5iDoN1BsrFr4kJ0bP8ayM_Q_ftEtyeb_FY2I=")) {
             writeUtf8("hi")
         }
@@ -79,30 +88,30 @@ class OkioMediaStoreTest {
 
     @Test
     fun shouldGetMediaWhenFileNotExists() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         cut.getMedia("url1")?.toByteArray()?.decodeToString() shouldBe null
     }
 
     @Test
     fun shouldDeleteMedia() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         fileSystem.write(basePath.resolve("K5pAaUF5iDoN1BsrFr4kJ0bP8ayM_Q_ftEtyeb_FY2I=")) {
             writeUtf8("hi")
         }
         cut.deleteMedia("url1")
-        fileSystem.listOrNull(basePath)?.size shouldBe 1
+        fileSystem.listOrNull(basePath)?.size shouldBe 2
     }
 
     @Test
     fun shouldDeleteMediaWhenFileNotExists() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         cut.deleteMedia("url1")
-        fileSystem.listOrNull(basePath)?.size shouldBe 1
+        fileSystem.listOrNull(basePath)?.size shouldBe 2
     }
 
     @Test
     fun shouldChangeMediaUrl() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         fileSystem.write(basePath.resolve("K5pAaUF5iDoN1BsrFr4kJ0bP8ayM_Q_ftEtyeb_FY2I=")) {
             writeUtf8("hi")
         }
@@ -112,8 +121,61 @@ class OkioMediaStoreTest {
 
     @Test
     fun shouldChangeMediaUrlWhenFileNotExists() = runTest {
-        cut.init()
+        cut.init(coroutineScope)
         cut.changeMediaUrl("url1", "url2")
-        fileSystem.listOrNull(basePath)?.size shouldBe 1
+        fileSystem.listOrNull(basePath)?.size shouldBe 2
+    }
+
+    @Test
+    fun shouldDeleteTmpDirectoryOnStartup() = runTest {
+        fileSystem.createDirectories(tmpPath)
+        fileSystem.write(tmpPath.resolve("tmp_file_1")) { writeUtf8("hi") }
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 1
+        cut.init(coroutineScope)
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 0
+    }
+
+    @Test
+    fun shouldDeleteTmpDirectoryOnShutdown() = runTest {
+        cut.init(coroutineScope)
+        fileSystem.write(tmpPath.resolve("tmp_file_1")) { writeUtf8("hi") }
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 1
+        coroutineScope.cancel()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe null
+    }
+
+    @Test
+    fun shouldCreateTemporaryFile() = runTest {
+        cut.init(coroutineScope)
+        cut.addMedia("url1", "hi".encodeToByteArray().toByteArrayFlow())
+        val platformMedia = cut.getMedia("url1").shouldNotBeNull()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 0
+        val tmpFile = platformMedia.getTemporaryFile().getOrThrow()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 1
+        fileSystem.read(tmpFile.path) { readUtf8() } shouldBe "hi"
+    }
+
+    @Test
+    fun shouldCreateTransformedTemporaryFile() = runTest {
+        cut.init(coroutineScope)
+        cut.addMedia("url1", "hi".encodeToByteArray().toByteArrayFlow())
+        val platformMedia = cut.getMedia("url1").shouldNotBeNull()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 0
+        val tmpFile = platformMedia
+            .transformByteArrayFlow { "###encrypted###".encodeToByteArray().toByteArrayFlow() }
+            .getTemporaryFile()
+            .getOrThrow()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 1
+        fileSystem.read(tmpFile.path) { readUtf8() } shouldBe "###encrypted###"
+    }
+
+    @Test
+    fun shouldDeleteTemporaryFile() = runTest {
+        cut.init(coroutineScope)
+        cut.addMedia("url1", "hi".encodeToByteArray().toByteArrayFlow())
+        val platformMedia = cut.getMedia("url1").shouldNotBeNull()
+        val tmpFile = platformMedia.getTemporaryFile().getOrThrow()
+        tmpFile.delete()
+        fileSystem.listOrNull(tmpPath)?.size shouldBe 0
     }
 }